feat(cart_builder,spreadsheet): add prototype weird layout height outside margin stretch scroll bug
